fix: change field ID fro uint to uint64 in the user model

+++ backend/src/repositories/user.go
@@ -17,7 +17,7 @@ func NewUsersRepositorie(db *sql.DB) *Users {
 }
 
 // Create insert a user in database
-func (repository Users) Create(user models.User) (ID uint, err error) {
+func (repository Users) Create(user models.User) (ID uint64, err error) {
 	stmt, err := repository.db.Prepare(
 		"INSERT INTO users (name, userName, email, password) VALUES ($1, $2, $3, $4) RETURNING id",
 	)
@@ -65,7 +65,7 @@ func (repository Users) FilterByUserName(userName string) (users []models.User, 
 }
 
 // GetById fetch an user by id
-func (repository Users) GetById(id int64) (user models.User, err error) {
+func (repository Users) GetById(id uint) (user models.User, err error) {
 	row, err := repository.db.Query(
 		"SELECT id, name, userName, email, creationDate FROM users WHERE id=$1", id,
 	)
